requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
future==0.18.3
h11==0.14.0
idna==3.4
openpyxl==3.1.0
outcome==1.2.0
pefile==2022.5.30
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PySocks==1.7.1
pywin32-ctypes==0.2.0
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
